Update main func
