+++ pkg/boltbrowser/boltbrowser.go
@@ -1,0 +1,9 @@
+package boltbrowser
+
+import internal "github.com/br0xen/boltbrowser/internal/boltbrowser"
+
+type Args = internal.Args
+
+var DefaultArgs = internal.DefaultArgs
+
+var Main = internal.Main
